broken, im a bit overwhelmed, taking a break
basically i am trying to make it so that magazines can store a value that persists across becoming an inventory object and back again. As of right now, when you pick up an inventory object all the static info is stored, ie name, stack size, icon, etc, and the amount is calculated based on how many times the method is called.