More work fixing magazine reloading functionality
mostly worked on "round i the chamber" mechanics, meaning a single bullet object is deducted from the magazine and added to a separate list. I had some issues with storing the info from an itemstat locally without creating a reference, so i did the the hard way value by value unfortunately. Still a lot of work to be done but you can pick up an empty gun and a magazine, hit reload, automatically find a magazine that fits within your inventory, load it in your gun, and load a round into the chamber. this actually creates an arror technically since the slide doesnt work like that on the type 100 but most guns will fucntion that way, shoulc create a exception later. also it doesnt work at all beyond that lol but still WIP